requirements.txt
llama-index==0.11.20
python-dotenv==1.0.1
llama-index-llms-ollama==0.3.6
llama-index-embeddings-huggingface==0.3.1
llama-index-readers-database==0.2.0
llama-parse==0.5.12
Flask==3.1.0